requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg==3.1.17
django-debug-toolbar